Move certification test creation to appropriate controller
