+++ api/lib/application/certification-courses/certification-course-controller.js
@@ -2,6 +2,7 @@ const certificationService = require('../../domain/services/certification-servic
 const certificationCourseService = require('../../../lib/domain/services/certification-course-service');
 const certificationSerializer = require('../../infrastructure/serializers/jsonapi/certification-serializer');
 const certificationCourseSerializer = require('../../infrastructure/serializers/jsonapi/certification-course-serializer');
+const usecases = require('../../domain/usecases');
 
 module.exports = {
 
@@ -22,4 +23,16 @@ module.exports = {
       .then((certificationCourse) => certificationCourseService.update(certificationCourse))
       .then(certificationSerializer.serializeFromCertificationCourse);
   },
+
+  save(request, h) {
+    const userId = request.auth.credentials.userId;
+    const accessCode = request.payload.data.attributes['access-code'];
+
+    return usecases.retrieveLastOrCreateCertificationCourse({ accessCode, userId })
+      .then(({ created, certificationCourse }) => {
+        const serialized = certificationCourseSerializer.serialize(certificationCourse);
+
+        return created ? h.response(serialized).created() : serialized;
+      });
+  }
 };
