+++ api/lib/application/certification-courses/index.js
@@ -39,8 +39,19 @@ exports.register = async function(server) {
         handler: certificationCourseController.update,
         tags: ['api']
       }
-    }
-
+    }, {
+      method: 'POST',
+      path: '/api/certification-courses',
+      config: {
+        handler: certificationCourseController.save,
+        notes: [
+          '- **Route nécessitant une authentification**\n' +
+          '- S\'il existe déjà une certification pour l\'utilisateur courant dans cette session, alors cette route renvoie la certification existante avec un code 200\n' +
+          '- Sinon, crée une certification pour l\'utilisateur courant dans la session indiquée par l\'attribut *access-code*, et la renvoie avec un code 201\n'
+        ],
+        tags: ['api']
+      }
+    },
   ]);
 };
