+++ api/tests/unit/application/certification-courses/index_test.js
@@ -1,7 +1,7 @@
 const { expect, sinon } = require('../../../test-helper');
 const Hapi = require('hapi');
 const securityController = require('../../../../lib/interfaces/controllers/security-controller');
-const certificationCoursesController = require('../../../../lib/application/certificationCourses/certification-course-controller');
+const certificationCoursesController = require('../../../../lib/application/certification-courses/certification-course-controller');
 
 describe('Unit | Application | Certifications Course | Route', function() {
 
@@ -13,10 +13,11 @@ describe('Unit | Application | Certifications Course | Route', function() {
     sinon.stub(certificationCoursesController, 'getResult').returns('ok');
     sinon.stub(certificationCoursesController, 'update').returns('ok');
     sinon.stub(certificationCoursesController, 'computeResult').returns('ok');
+    sinon.stub(certificationCoursesController, 'save').returns('ok');
 
     server = Hapi.server();
 
-    return server.register(require('../../../../lib/application/certificationCourses'));
+    return server.register(require('../../../../lib/application/certification-courses'));
   });
 
   describe('GET /api/admin/certifications/{id}/details', () => {
@@ -68,6 +69,25 @@ describe('Unit | Application | Certifications Course | Route', function() {
 
       // then
       expect(response.statusCode).to.equal(200);
+    });
+  });
+
+  describe('POST /api/certification-courses', () => {
+
+    it('should exist', () => {
+      // given
+      const options = {
+        method: 'POST',
+        url: '/api/certification-courses'
+      };
+
+      // when
+      const promise = server.inject(options);
+
+      // then
+      return promise.then((res) => {
+        expect(res.statusCode).to.equal(200);
+      });
     });
   });
 });
